+++ src/DTO/Payment.php
@@ -5,8 +5,10 @@ declare(strict_types=1);
 namespace Isapp\CashierSupport\DTO;
 
 use Carbon\CarbonImmutable;
-use Isapp\CashierSupport\Enums\Currency;
+use Isapp\CashierSupport\Casts\CurrencyCast;
 use Isapp\CashierSupport\Enums\PaymentStatus;
+use Money\Currency;
+use Spatie\LaravelData\Attributes\WithCastAndTransformer;
 use Spatie\LaravelData\Data;
 
 /**
@@ -20,6 +22,7 @@ class Payment extends Data
     public function __construct(
         public string $id,
         public int $amount,
+        #[WithCastAndTransformer(CurrencyCast::class)]
         public Currency $currency,
         public PaymentStatus $status,
         public ?string $paymentMethodId = null,